the reducer is divided into components, the filtering of films has been moved to select
