+++ package.json
@@ -23,6 +23,7 @@
     "eslint-plugin-react": "7.12.4",
     "jest": "24.8.0",
     "react-test-renderer": "16.8.6",
+    "redux-logger": "3.0.6",
     "webpack": "4.30.0",
     "webpack-cli": "3.3.1",
     "webpack-dev-server": "3.3.1"
@@ -38,6 +39,7 @@
     "react-redux": "7.0.3",
     "recompose": "0.30.0",
     "redux": "4.0.1",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
