+++ src/index.jsx
@@ -4,16 +4,23 @@ import {Provider} from "react-redux";
 import {createStore, applyMiddleware} from "redux";
 import thunk from "redux-thunk";
 import {compose} from "recompose";
+import logger from 'redux-logger';
 
 import App from "./components/app/app.jsx";
-import {reducer, Operation} from "./reducer.js";
-import api from "./api.js";
+import reducer from "./reducer/index.js";
+import {Operation} from "./reducer/movie/movie.js";
+import createAPI from "./api.js";
 
+
+const api = createAPI((...args) => store.dispatch(...args));
 
 const store = createStore(
     reducer,
     compose(
-        applyMiddleware(thunk.withExtraArgument({api})),
+        applyMiddleware(
+            logger,
+            thunk.withExtraArgument(api)
+        ),
         window && window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
     )
 );
