+++ src/reducer/movie/selectors.js
@@ -1,0 +1,25 @@
+import {createSelector} from "reselect";
+
+
+const getGenre = (state) => (
+  state.movie.genre
+);
+
+const getFilms = (state) => (
+  state.movie.films
+);
+
+export const getFilterFilms = createSelector(
+    getGenre,
+
+    getFilms,
+
+    (selectedGenre, filmsList) => {
+      if (selectedGenre === `All genres`) {
+        return filmsList;
+      }
+      return filmsList.filter((film) =>
+        film.genre.some((filmGenre) =>
+          filmGenre === selectedGenre));
+    }
+);
